techpack: touch: goodix: Remove WQ_CPU_INTENSIVE from unbound workqueue allocations
WQ_CPU_INTENSIVE is useful for bound work items which are expected to hog CPU
cycles so that their execution is regulated by the system scheduler and thus
do not contribute to the concurrency level, which is pointless for an unbound
workqueue.
Therefore, using WQ_CPU_INTENSIVE with WQ_UNBOUND is meaningless. Remove it.

Documentation: https://www.kernel.org/doc/html/latest/core-api/workqueue.html#flags

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>